Tachyos v0.0.2 1

+++ include/kernel/tch_fs.h
@@ -25,7 +25,7 @@ struct tch_file_operations {
 	size_t (*read)(struct tch_file* filp,char* rdata,size_t len);
 	size_t (*write)(struct tch_file* filp,const char* wdata,size_t len);
 	int (*close)(struct tch_file* filp);
-	off_t (*seek)(struct tch_file* filp,off_t offset,int whence);
+	size_t (*seek)(struct tch_file* filp,size_t offset,int whence);
 };
